+++ src/test/java/com/api/utils/CSVReaderUtil.java
@@ -1,0 +1,44 @@
+package com.api.utils;
+
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.List;
+
+import com.dataproviders.api.bean.UserBean;
+import com.opencsv.CSVReader;
+import com.opencsv.bean.CsvToBean;
+import com.opencsv.bean.CsvToBeanBuilder;
+
+public class CSVReaderUtil {
+
+	
+	/*
+	 * 
+	 * Constuctor is private
+	 * satic methods
+	 * 
+	 * Job:- Help me read the CSV file and Map it a Bean
+	 */
+	
+	private CSVReaderUtil() {
+		
+	}
+	
+	public static void loadCSV(String pathOfCSVFile) {
+
+			InputStream is = Thread.currentThread().getContextClassLoader().getResourceAsStream(pathOfCSVFile);
+			InputStreamReader isr = new InputStreamReader(is);
+			CSVReader csvReader = new CSVReader(isr);
+
+			CsvToBean<UserBean> csvToBean = new CsvToBeanBuilder(csvReader)
+					.withType(UserBean.class)
+					.withIgnoreEmptyLine(true)
+					.build();
+
+			List<UserBean> userList = csvToBean.parse();	
+			System.out.println(userList);
+
+		}
+	}
+	
+
