+++ src/test/java/com/dataproviders/api/bean/UserBean.java
@@ -1,18 +1,12 @@
-package com.demo.csv;
+package com.dataproviders.api.bean;
 
-public class UserPOJO {
+public class UserBean {
 	
 	private String username;
 	private String password;
 	
-	public UserPOJO() {
+	public UserBean() {
 		
-	}
-	
-	public UserPOJO(String username, String password) {
-		super();
-		this.username = username;
-		this.password = password;
 	}
 
 	public String getUsername() {
